Rename packages to be lowercase
