+++ app/src/androidTest/java/com/duckduckgo/app/di/TestAppComponent.kt
@@ -17,7 +17,7 @@
 package com.duckduckgo.app.di
 
 import android.app.Application
-import com.duckduckgo.app.browser.autoComplete.BrowserAutoCompleteModule
+import com.duckduckgo.app.browser.autocomplete.BrowserAutoCompleteModule
 import com.duckduckgo.app.browser.di.BrowserModule
 import com.duckduckgo.app.browser.favicon.FaviconModule
 import com.duckduckgo.app.browser.rating.di.RatingModule
