+++ app/src/main/AndroidManifest.xml
@@ -129,7 +129,7 @@
             android:name="com.duckduckgo.app.bookmarks.ui.BookmarksActivity"
             android:label="@string/bookmarksActivityTitle" />
         <activity
-            android:name=".defaultBrowsing.DefaultBrowserInfoActivity"
+            android:name=".defaultbrowsing.DefaultBrowserInfoActivity"
             android:theme="@style/ModalCardTheme" />
         <activity
             android:name="com.duckduckgo.app.fire.FireActivity"
