+++ app/src/main/java/com/duckduckgo/app/di/AndroidBindingModule.kt
@@ -20,7 +20,7 @@ import com.duckduckgo.app.about.AboutDuckDuckGoActivity
 import com.duckduckgo.app.bookmarks.ui.BookmarksActivity
 import com.duckduckgo.app.browser.BrowserActivity
 import com.duckduckgo.app.browser.BrowserTabFragment
-import com.duckduckgo.app.browser.defaultBrowsing.DefaultBrowserInfoActivity
+import com.duckduckgo.app.browser.defaultbrowsing.DefaultBrowserInfoActivity
 import com.duckduckgo.app.browser.rating.ui.AppEnjoymentDialogFragment
 import com.duckduckgo.app.browser.rating.ui.GiveFeedbackDialogFragment
 import com.duckduckgo.app.browser.rating.ui.RateAppDialogFragment
