+++ app/src/main/java/com/duckduckgo/app/di/AppComponent.kt
@@ -18,7 +18,7 @@ package com.duckduckgo.app.di
 
 
 import android.app.Application
-import com.duckduckgo.app.browser.autoComplete.BrowserAutoCompleteModule
+import com.duckduckgo.app.browser.autocomplete.BrowserAutoCompleteModule
 import com.duckduckgo.app.browser.di.BrowserModule
 import com.duckduckgo.app.browser.favicon.FaviconModule
 import com.duckduckgo.app.browser.rating.di.RatingModule
