+++ app/src/main/res/layout-land/activity_default_browser_info.xml
@@ -21,7 +21,7 @@
     android:layout_height="match_parent"
     android:layout_margin="30dp"
     android:background="@drawable/notification_info_bottom_background"
-    tools:context="com.duckduckgo.app.browser.defaultBrowsing.DefaultBrowserInfoActivity">
+    tools:context="com.duckduckgo.app.browser.defaultbrowsing.DefaultBrowserInfoActivity">
 
     <View
         android:id="@+id/titleSection"
